Refactor `Dockerfile`
Changed to using `alpine` for the base image, and split into a
multi-stage build. The image size is now only 120MB, down from 765MB. One
thing that needs to be looked into more is the output `/bin/sh: fmt: not
found` during the `fteqw` compilation step. Nothing seems to be breaking
(on the surface) when playtesting, but should review.